Add support for opening a column family with options
Adds a new `DB::open_cf_descriptors` method that allows passing in Options for
each column family.  A new `ColumnFamilyDescriptor` type was added to contain
the congfiguration for each column family. This fixes an issue where a column
family was created with non-default options, and then fails to re-open due to a
config mismatch.